+++ src/main.tsx
@@ -5,14 +5,15 @@ import "@github/spark/spark"
 
 import App from './App.tsx'
 import { ErrorFallback } from './ErrorFallback.tsx'
-import { A11yLiveRegion } from './components/A11yLiveRegion'
+import { A11yProvider } from './components/A11yLiveRegion'
 
 import "./main.css"
 
 createRoot(document.getElementById('root')!).render(
   <ErrorBoundary FallbackComponent={ErrorFallback}>
-    <App />
-    <A11yLiveRegion />
+    <A11yProvider>
+      <App />
+    </A11yProvider>
     <Toaster
       position="top-center"
       richColors